refactor(connlib): introduce otel-instruments crate for meters
Move all OpenTelemetry instrument definitions recorded on the 'connlib'
meter into a dedicated crate, so each instrument has a single, consistent
definition and call sites differ only by their attributes. Consumers
depend on the crate directly. The periodic queue-length gauge helper and
the QueueLength trait move there too.
