+++ rust/client-ffi/Cargo.toml
@@ -20,6 +20,7 @@ itertools = { workspace = true }
 libc = { workspace = true }
 logging = { workspace = true }
 opentelemetry = { workspace = true }
+otel-instruments = { workspace = true }
 phoenix-channel = { workspace = true }
 rustls = { workspace = true }
 secrecy = { workspace = true }
