+++ rust/libs/connlib/bufferpool/Cargo.toml
@@ -11,6 +11,7 @@ path = "lib.rs"
 bytes = { workspace = true }
 crossbeam-queue = { workspace = true }
 opentelemetry = { workspace = true, features = ["metrics"] }
+otel-instruments = { workspace = true }
 
 [dev-dependencies]
 opentelemetry_sdk = { workspace = true, features = ["testing", "metrics"] }
