+++ rust/libs/connlib/otel-instruments/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "eventloop-budget"
+name = "otel-instruments"
 version = "0.1.0"
 edition = { workspace = true }
 license = { workspace = true }
 
 [dependencies]
 opentelemetry = { workspace = true, features = ["metrics"] }
-otel-instruments = { workspace = true }
+tokio = { workspace = true, features = ["time", "sync"] }
 
 [lints]
 workspace = true
