+++ rust/libs/connlib/snownet/Cargo.toml
@@ -19,6 +19,7 @@ is = { workspace = true }
 itertools = { workspace = true }
 logging = { workspace = true }
 opentelemetry = { workspace = true, features = ["metrics"] }
+otel-instruments = { workspace = true }
 rand = { workspace = true }
 ringbuffer = { workspace = true }
 smallvec = { workspace = true }
