+++ rust/libs/connlib/tunnel/Cargo.toml
@@ -39,6 +39,7 @@ l4-udp-dns-server = { workspace = true }
 logging = { workspace = true }
 lru = { workspace = true }
 opentelemetry = { workspace = true, features = ["metrics"] }
+otel-instruments = { workspace = true }
 proptest = { workspace = true, optional = true }
 rand = { workspace = true }
 rangemap = { workspace = true }
