+++ rust/libs/connlib/tunnel/src/otel.rs
@@ -64,10 +64,6 @@ pub mod attr {
     }
 }
 
-pub mod metrics {
-    pub use telemetry::otel::metrics::*;
-}
-
 #[cfg(test)]
 mod tests {
     use super::attr::dns_question_type;
