+++ rust/libs/telemetry/src/otel.rs
@@ -501,113 +501,6 @@ mod normalize_tests {
     }
 }
 
-pub mod metrics {
-    use std::{ops::ControlFlow, time::Duration};
-
-    use opentelemetry::{
-        KeyValue,
-        metrics::{Counter, Gauge, Histogram},
-    };
-
-    use crate::otel::QueueLength;
-
-    pub fn network_packet_dropped() -> Counter<u64> {
-        opentelemetry::global::meter("connlib")
-            .u64_counter("network.packet.dropped")
-            .with_description("Count of packets that are dropped or discarded")
-            .with_unit("{packet}")
-            .build()
-    }
-
-    pub fn tunnel_errors() -> Counter<u64> {
-        opentelemetry::global::meter("connlib")
-            .u64_counter("tunnel.error")
-            .with_description("Number of errors encountered while processing a packet batch.")
-            .with_unit("{error}")
-            .build()
-    }
-
-    pub fn portal_connection_hiccups() -> Counter<u64> {
-        opentelemetry::global::meter("connlib")
-            .u64_counter("portal.connection.hiccup")
-            .with_description("Number of portal connection hiccups by cause.")
-            .with_unit("{hiccup}")
-            .build()
-    }
-
-    pub fn connection_count() -> Gauge<u64> {
-        opentelemetry::global::meter("connlib")
-            .u64_gauge("tunnel.connection.count")
-            .with_description("Number of connections by the network path in use.")
-            .with_unit("{connection}")
-            .build()
-    }
-
-    /// Measures how long connlib takes to recursively resolve a DNS query against an upstream resolver.
-    pub fn dns_lookup_duration() -> Histogram<f64> {
-        opentelemetry::global::meter("connlib")
-            .f64_histogram("dns.lookup.duration")
-            .with_description("Duration of a recursive DNS lookup against an upstream resolver.")
-            .with_unit("s")
-            .with_boundaries(vec![
-                0.005, 0.01, 0.025, 0.05, 0.075, 0.1, 0.25, 0.5, 0.75, 1.0, 2.5, 5.0, 7.5, 10.0,
-            ])
-            .build()
-    }
-
-    pub async fn periodic_system_queue_length<const N: usize>(
-        queue: impl QueueLength,
-        attributes: [KeyValue; N],
-    ) {
-        let gauge = opentelemetry::global::meter("connlib")
-            .u64_gauge("system.queue.length")
-            .with_description("The length of a queue.")
-            .build();
-
-        periodic_gauge(
-            gauge,
-            |gauge| {
-                let len = match queue.queue_length() {
-                    Some(len) => len,
-                    None => return ControlFlow::Break(()),
-                };
-
-                gauge.record(len, &attributes);
-
-                ControlFlow::Continue(())
-            },
-            Duration::from_secs(1),
-        )
-        .await;
-    }
-
-    pub async fn periodic_gauge<T>(
-        gauge: Gauge<T>,
-        callback: impl Fn(&Gauge<T>) -> ControlFlow<(), ()>,
-        interval: Duration,
-    ) {
-        while callback(&gauge).is_continue() {
-            tokio::time::sleep(interval).await;
-        }
-    }
-}
-
-pub trait QueueLength: Send + Sync + 'static {
-    fn queue_length(&self) -> Option<u64>;
-}
-
-impl<T> QueueLength for tokio::sync::mpsc::WeakSender<T>
-where
-    T: Send + Sync + 'static,
-{
-    fn queue_length(&self) -> Option<u64> {
-        let sender = self.upgrade()?;
-        let len = sender.max_capacity() - sender.capacity();
-
-        Some(len as u64)
-    }
-}
-
 pub fn default_resource_with<const N: usize>(attributes: [KeyValue; N]) -> Resource {
     Resource::builder_empty()
         .with_detector(Box::new(TelemetryResourceDetector))
